Remove cached files
