+++ ascii_art/__init__.py
@@ -1,0 +1,24 @@
+from .converter import (
+    char_array,
+    get_char,
+    load_char_array,
+    list_files_from_assets,
+    convert_image,
+    convert_video,
+    loader,
+    print_divider,
+)
+from .cli import parse_args, main
+
+__all__ = [
+    "char_array",
+    "get_char",
+    "load_char_array",
+    "list_files_from_assets",
+    "convert_image",
+    "convert_video",
+    "loader",
+    "print_divider",
+    "parse_args",
+    "main",
+]
